requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
streamlit==1.32.0
langchain_groq
pypdf
faiss-cpu
langchain_huggingface
